Implement new Graph

+++ src/HamiltonVisualizer/Constants/ConstantValues.cs
@@ -29,12 +29,12 @@ namespace HamiltonVisualizer.Constants
 
         public static class ControlSpecifications
         {
-            public const int NodeWidth = 34;
+            public const double NodeWidth = 34;
 
             /// <summary>
             /// The drawing canvas width and height.
             /// </summary>
-            public const int DrawingCanvasSidesLength = 1440;
+            public const double DrawingCanvasSidesLength = 1440;
         }
 
         public static class Time
